+++ dev/console/console.js
@@ -226,6 +226,90 @@ var CKCONSOLE = ( function() {
 		logItemTpl = new CKEDITOR.template(
 			'<li class="ckconsole_log_item ckconsole_fresh_log_item"><time datetime="{time}" class="ckconsole_time">{time}</time> <code>{msg}</code></li>'
 		);
-
 	return that;
 } )();
+
+/**
+ * Usage:
+ *
+ * CKCONSOLE.addEventPanel( 'drop' );
+ * CKCONSOLE.create( 'drop', { editor: 'inline' } );
+ *
+ * or (with observed types):
+ *
+ * CKCONSOLE.addEventPanel( 'drop', [ 'nativeEvent', 'target' ] );
+ * CKCONSOLE.create( 'drop', { editor: 'inline' } );
+ *
+ * or (with observed types mapping):
+ *
+ * CKCONSOLE.addEventPanel( 'drop',
+ * 		[ 'nativeEvent', 'target' ],
+ * 		function( evt ) {
+ * 			return {
+ * 				'nativeEvent': evt.data.nativeEvent,
+ * 				'target': evt.data.target
+ * 			};
+ * 		}
+ * );
+ * CKCONSOLE.create( 'drop', { editor: 'inline' } );
+ *
+ */
+CKCONSOLE.addEventPanel = function( evtName, types, mapping ) {
+	var values,
+		content = '<ul class="ckconsole_list">';
+
+	if ( !types ) {
+		var types = [];
+	}
+	if ( !mapping ) {
+			mapping = function( evt ) {
+				var ret = {};
+
+				for ( var i = 0; i < types.length; i++ ) {
+					ret[ types[ i ] ] = evt.data[ types[ i ] ];
+				};
+				return ret;
+			}
+	}
+
+	for ( var i = 0; i < types.length; i++ ) {
+		content += '<li>'+ types[ i ] + ': <span class="ckconsole_value" data-value="'+ types[ i ] + '"></span></li>';
+	};
+	content += '</ul>';
+
+	CKCONSOLE.add( evtName, {
+		panels: [
+			{
+				type: 'box',
+				content: content,
+
+				refresh: function( editor ) {
+					values.header = evtName;
+					return values;
+				},
+
+				refreshOn: function( editor, refresh ) {
+					editor.on( evtName, function( evt ) {
+						values = mapping( evt );
+						refresh();
+					} );
+				}
+			},
+			{
+				type: 'log',
+				on: function( editor, log, logFn ) {
+					editor.on( evtName, function( evt ) {
+						var logStr = evtName + '; '
+						values = mapping( evt );
+
+						for ( var i = 0; i < types.length; i++ ) {
+							logStr += types[ i ] + ': ' + values[ types[ i ] ] +'; ';
+						}
+
+						logFn( logStr )();
+					} );
+				}
+			}
+		]
+	} );
+}